Transportation review and correction
corrected the previous code and made it  more cleaner

+++ app/Services/Transportation/TransportationModelService.php
@@ -1,16 +1,22 @@
 <?php
 
 namespace App\Services\Transportation;
+
+
 use App\Services\BaseService;
 use App\Repositories\Transportations\TransportationModelRepository;
+
+use App\Repositories\Transportations\TransportationCategoryRepository;
 
 class TransportationModelService extends BaseService
 {
     public $TransportationModel;
-    public function __construct()
+    public $TransportationMode;
+    public function __construct(TransportationModelRepository $transportationModelRepository, TransportationCategoryRepository $transportationMode)
     {
-        parent::__construct();
-        $this->TransportationModel = new TransportationModelRepository();
+
+        $this->TransportationModel = $transportationModelRepository;
+        $this->TransportationMode = $transportationMode;
     }
 
     public function All()
@@ -21,6 +27,14 @@ class TransportationModelService extends BaseService
 
     public function create($reqData)
     {
+        $categoryExist = $this->TransportationMode->findbyId($reqData['category_id']);
+        if (blank($categoryExist)) {
+            abort(401, "category Id  not Found");
+        }
+        $nameExist = $this->TransportationModel->findData("name", $reqData['name']);
+        if ($nameExist) {
+            abort(401, "Transportation Model Name Exist");
+        }
         $resp = $this->TransportationModel->create($reqData);
         return $resp;
     }
@@ -28,17 +42,30 @@ class TransportationModelService extends BaseService
     public function findbyId($id)
     {
         $resp = $this->TransportationModel->findById($id);
+
+        if (blank($resp)) {
+            abort(401, "Transportation model not found");
+        }
         return $resp;
     }
 
     public function update($reqData, $id)
     {
+        $checkId = $this->findbyId($id);
+        if (blank($checkId)) {
+            abort(401, "Transportation model not found");
+        }
         $resp = $this->TransportationModel->update($reqData, $id);
         return $resp;
     }
 
-    public function delete($id){
-        $resp =$this->TransportationModel->delete($id);
+    public function delete($id)
+    {
+        $checkId = $this->findbyId($id);
+        if (blank($checkId)) {
+            abort(401, "Transportation model not found");
+        }
+        $resp = $this->TransportationModel->delete($id);
         return $resp;
     }
 }
